+++ collector/src/main/java/com/bitmovin/analytics/CustomDataHelpers.kt
@@ -1,0 +1,12 @@
+package com.bitmovin.analytics
+
+import com.bitmovin.analytics.data.CustomData
+
+internal class CustomDataHelpers {
+    interface Getter {
+        fun getCustomData(): CustomData
+    }
+    interface Setter {
+        fun setCustomData(customData: CustomData)
+    }
+}