+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Development v2
 
+### Changed
+- BitmovinPlayer v3 collector behaviour when using `setCustomData` and `setCustomDataOnce` and custom source configs
+
 ## Development v1
 
 ## v1.24.0